Move tests utils from __test__
