+++ shell/edit/networking.k8s.io.networkpolicy/__tests__/utils/mock.json
